With new storeClientCopy() scheme, the client side might re-request some data which has already been read from disk (if it didn't get the full HTTP headers).  That means we need to put lseek() back in; ugh.  New fde->disk.offset tracks the current file offset.
Fixed the 'double-read' bug found by Alex Rousskov.  Always
make the read callback after each read() call.  Don't wait
for the read buffer to full or for EOF.

This means dread_ctrl->offset isn't needed as it was being used
because there will only be one read() per dread_ctrl.  ->offset
now becomes the offset requested from the store module, and we
seek if dread_ctrl->offset != fde->disk.offset.

+++ src/store.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: store.cc,v 1.288 1997/10/17 04:00:12 wessels Exp $
+ * $Id: store.cc,v 1.289 1997/10/17 05:20:19 wessels Exp $
  *
  * DEBUG: section 20    Storeage Manager
  * AUTHOR: Harvest Derived
@@ -451,7 +451,7 @@ storeLog(int tag, const StoreEntry * e)
 	(int) reply->expires,
 	reply->content_type[0] ? reply->content_type : "unknown",
 	reply->content_length,
-	mem->inmem_hi - mem->reply->hdr_sz,
+	(int) (mem->inmem_hi - mem->reply->hdr_sz),
 	RequestMethodStr[e->method],
 	mem->log_url);
     file_write(storelog_fd,
